+++ src/typings/config.d.ts
@@ -21,6 +21,7 @@ declare module '@typings' {
     timeline_path: string;
     enable_top_memories: boolean;
     stack_raw_files: boolean;
+    dedup_identical: boolean;
 
     // viewer settings
     high_res_cond_default: HighResCond;